+++ Excel_PRIME/XlsbImp/XlsbWorkBookReader.cs
@@ -255,7 +255,8 @@ internal sealed class XlsbWorkBookReader : IOpenXmlWorkBookReaderAsync
                     int row = nextRecord.GetInt32(offset) + 1;
                     offset += 4;
                     int col = nextRecord.GetInt16(offset) + 1;
-                    return (col.GetExcelColumnName(), col.GetExcelColumnName(), row, row, false, sheetRef);
+                    string excelColumnName = col.GetExcelColumnName();
+                    return (excelColumnName, excelColumnName, row, row, false, sheetRef);
                 }
 
             case 0x3B:  //PtgArea3d
@@ -282,7 +283,8 @@ internal sealed class XlsbWorkBookReader : IOpenXmlWorkBookReaderAsync
         //    sheetRef = nextRecord.GetInt16(offset + 5);
         //}
 
-        return (colFirst.GetExcelColumnName(), colLast.GetExcelColumnName(), rowFirst, rowLast, false, sheetRef);
+        string columnName = colLast.GetExcelColumnName();
+        return (colFirst.GetExcelColumnName(), columnName, rowFirst, rowLast, false, sheetRef);
     }
 
     private void Dispose(bool isDisposing)